Move parse_size to holland.lib.mysql
